[OG-74] Equipment
