+++ polymorphia-frontend/app/(logged-in)/equipment/index.css
@@ -1,0 +1,6 @@
+@import "tailwindcss";
+@import "../../../styles/globals.css";
+
+.equipment {
+    @apply w-full md:max-w-5xl mx-auto p-4 opacity-100 select-none;
+}
