+++ polymorphia-frontend/app/(logged-in)/equipment/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import "../../../styles/equipment.css";
+import "./index.css";
 import { useTitle } from "@/components/navigation/TitleContext";
 import { useEffect } from "react";
 import ItemModal from "@/components/equipment/modals/ItemModal";
@@ -7,10 +7,10 @@ import ChestModal from "@/components/equipment/modals/ChestModal";
 
 import { useQuery } from "@tanstack/react-query";
 import Loading from "@/components/general/Loading";
-import EquipmentService from "@/services/EquipmentService";
+import Service from "@/app/(logged-in)/equipment/service";
 import EquipmentSectionWrapper from "@/components/equipment/EquipmentSectionWrapper";
 import OpeningChestModal from "@/components/equipment/modals/OpeningChestModal";
-import { EquipmentProvider } from "@/components/providers/EquipmentContext";
+import { EquipmentProvider } from "@/components/providers/equipment/EquipmentContext";
 
 export default function Equipment() {
   const { setTitle } = useTitle();
@@ -22,12 +22,12 @@ export default function Equipment() {
   // TODO: error handling
   const { data: items, isLoading: isItemsLoading } = useQuery({
     queryKey: ["equipment-items"],
-    queryFn: () => EquipmentService.getItems(),
+    queryFn: () => Service.getItems(),
   });
 
   const { data: chests, isLoading: isChestsLoading } = useQuery({
     queryKey: ["equipment-chests"],
-    queryFn: () => EquipmentService.getChests(),
+    queryFn: () => Service.getChests(),
   });
 
   if (isItemsLoading || isChestsLoading) {
