+++ polymorphia-frontend/components/equipment/index.css
@@ -1,0 +1,62 @@
+@import "tailwindcss";
+@import "../../styles/globals.css";
+
+/* EquipmentSection */
+
+.equipment-header {
+    @apply text-6xl mb-6 lg:mb-10;
+}
+
+.equipment-grid {
+    @apply grid grid-cols-2 sm:grid-cols-4 gap-4;
+}
+
+.equipment-grid-item {
+    @apply relative w-full aspect-square;
+}
+
+.equipment-image {
+    @apply object-cover rounded-2xl shadow-lg;
+}
+
+.equipment-image-badge {
+    @apply text-3xl w-10 rounded-tl-2xl rounded-br-2xl;
+}
+
+.equipment-chest-btn-wrapper {
+    @apply w-full flex-centered mt-4;
+}
+
+.equipment-chest-btn {
+    @apply w-full py-0 text-lg lg:text-2xl rounded-xl;
+}
+
+.equipment-locked-item {
+    @apply absolute bg-neutral-300/70 dark:bg-primary-dark/80 flex-centered w-full h-full;
+
+    span {
+        @apply material-symbols;
+
+        p {
+            @apply text-8xl opacity-80;
+        }
+    }
+}
+
+/* OpeningChestModalContent */
+
+.opening-chest-modal {
+    @apply w-80 grid grid-cols-2 gap-3;
+}
+
+.opening-chest-modal-image-wrapper {
+    @apply relative w-full aspect-square cursor-pointer;
+}
+
+.opening-chest-modal-image-selected {
+    @apply outline-4 outline-amber-400;
+}
+
+.opening-chest-modal-button-wrapper {
+    @apply w-full mt-5;
+}
