+++ polymorphia-frontend/components/equipment/types.ts
@@ -1,0 +1,40 @@
+import { ChestBehavior } from "@/interfaces/api/DTO";
+
+export interface EquipmentSectionWrapperProps {
+  items: ItemData[];
+  chests: ChestData[];
+}
+
+export type EquipmentSectionType = "item" | "chest";
+
+export interface EquipmentSectionProps {
+  type: EquipmentSectionType;
+  data: ItemData[] | ChestData[];
+}
+
+export interface ItemData {
+  itemId: number;
+  title: string;
+  subtitle: string;
+  imageUrl: string;
+  quantity: number;
+  items: Item[];
+}
+
+export interface ChestData {
+  chestId: number;
+  title: string;
+  subtitle: string;
+  imageUrl: string;
+  behavior: ChestBehavior;
+  openedDate: string | undefined;
+  items: Item[];
+}
+
+export interface Item {
+  itemId: number;
+  title: string;
+  subtitle: string;
+  imageUrl: string;
+  bonusXp?: string;
+}
